requirements.txt
-e .
pyyaml
colorlog
pyserial
observable